feat: extract variant list component
