+++ src/annotation_extraction/components.py
@@ -1,0 +1,43 @@
+from src.annotation_extraction.simple_inference import Variant, VariantList, SimpleLLM, PromptGenerator
+from src.annotation_extraction.article_parser import ArticleParser
+from loguru import logger
+import json
+
+VARIANT_LIST_PROMPT = """
+You are an expert pharmacogenomics researcher reading and extracting annotations from the following article:
+
+{article_text}
+
+From this article, note down ALL discussed variants/haplotypes (ex. rs113993960, CYP1A1*1, etc.) that are part of the article's
+study/analysis. Include information on the gene group and allele (if present). Your output format should be a list of the variants with the following attributes:
+Variant: The Variant / Haplotypes (ex. rs2909451, CYP2C19*1, CYP2C19*2, *1/*18, etc.)
+Gene: The gene group of the variant (ex. DPP4, CYP2C19, KCNJ11, etc.)
+Allele: Specific allele or genotype if different from variant (ex. TT, *1/*18, del/del, etc.)
+"""
+
+def extract_variants_list(article_text: str = None, pmcid: str = None) -> VariantList:
+    """Extract a list of variants from an article.
+    Args:
+        article_text: The text of the article.
+        PMCID: The PMCID of the article.
+
+    Returns:
+        A list of variants.
+    """
+    if article_text is None and pmcid is None:
+        logger.error("Either article_text or pmcid must be provided.")
+        raise ValueError("Either article_text or pmcid must be provided.")
+    
+    if article_text is not None and pmcid is not None:
+        logger.warning("Both article_text and PMCID are provided. Using article_text.")
+
+    if article_text is None:
+        article_text = ArticleParser(pmcid=pmcid).get_article_text()
+
+    model = SimpleLLM()
+    prompt_generator = PromptGenerator(VARIANT_LIST_PROMPT, {"article_text": article_text})
+    prompt = prompt_generator.get_prompt()
+    output = model.generate(prompt, response_format=VariantList)
+    variant_list = json.loads(output)['variant_list']
+    # variant_list_strs = [i['variant_id'] for i in variant_list]
+    return variant_list
